string serializers now handle empty strings /3

+++ protocol/RcpTypesExtensions.cs
@@ -37,7 +37,7 @@ namespace RCP.Protocol
                     writer.Write(bytes, 0, length);
                 }
                 else
-                    writer.Write((byte)0);
+                    writer.Write((ushort)0);
             }
         }
     }
@@ -56,7 +56,7 @@ namespace RCP.Protocol
                     writer.Write(bytes, 0, length);
                 }
                 else
-                    writer.Write((byte)0);
+                    writer.Write((int)0);
             }
         }
     }
